% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boost_lines.R
\name{bAppend}
\alias{bAppend}
\title{Concatenate a List of BoostLines Objects}
\usage{
bAppend(x)
}
\arguments{
\item{x}{A List of BoostLines objects.}
}
\value{
A BoostLine S3 object.
}
\description{
Creates a single BoostLines object from multiple BoostLines objects.
}

