% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boost_lines.R
\name{bDistance}
\alias{bDistance}
\title{Calculate BoostLines Minimal Distances}
\usage{
bDistance(bl)
}
\arguments{
\item{bl}{A BoostLines object.}
}
\value{
A numeric matrix.
}
\description{
Calculates minimal distances between BoostLines.
}

