% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boost_lines.R
\name{bGrid}
\alias{bGrid}
\title{Create a Gridded Line Collection}
\usage{
bGrid(x, nrow = NULL, ncol = NULL, res = NULL, n8 = TRUE,
  proj4string = CRS(""))
}
\arguments{
\item{x}{Upper left coordinate pair OR a RasterLayer object OR a VeloxRaster object.}

\item{nrow}{Number of rows. Required if x is a coordinate pair.}

\item{ncol}{Number of columns. Required if x is a coordinate pair.}

\item{res}{Resolution (x and y). Required if x is a coordinate pair.}

\item{n8}{Boolean. Whether 8 neighborhood should be connected. If FALSE, connects 4 neighborhood.}

\item{proj4string}{A proj4string. Ignored if x is a RasterLayer OR a VeloxRaster object.}
}
\value{
A BoostLine S3 object.
}
\description{
Creates a BoostLines object with gridded lines.
}

