% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boost_lines.R
\name{bIntersects}
\alias{bIntersects}
\title{Test BoostLines Intersections}
\usage{
bIntersects(bl)
}
\arguments{
\item{bl}{A BoostLines object.}
}
\value{
A logical matrix.
}
\description{
Checks whether BoostLines intersect.
}

