% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boost_lines.R
\name{bLength}
\alias{bLength}
\title{Calculate BoostLines Lengths}
\usage{
bLength(bl, lonlat)
}
\arguments{
\item{bl}{A BoostLines object.}

\item{lonlat}{A boolean; are line coordinates in lon/lat? If TRUE, returns lenghts in meters.}
}
\value{
A numeric vector.
}
\description{
Calculates line lengths.
}

