% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boost_lines.R
\name{bNode}
\alias{bNode}
\title{Node BoostLines object}
\usage{
bNode(bl)
}
\arguments{
\item{bl}{A BoostLines object.}
}
\value{
A list of (1) the noded BoostLines object and (2) an integer vector of line IDs.
}
\description{
Breaks lines at all proper intersections.
}

