% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boost_lines.R
\name{boost2graph}
\alias{boost2graph}
\title{Fast boost2graph}
\usage{
boost2graph(x, df, lonlat = FALSE, plot.result = FALSE)
}
\arguments{
\item{x}{A BoostLine object.}

\item{df}{A DataFrame with as many rows as there are lines in \code{x}.}

\item{lonlat}{A boolean indiciating whether BoostLines are in lat/lon format.
If TRUE, the length edge attribute added to the igraph object is measured in meters.}

\item{plot.result}{Should resulting graph be plotted?}
}
\value{
A \code{igraph::graph} object.
}
\description{
Generates an igraph graph from a BoostLine object and a DataFrame.
}
\details{
Generates an igraph graph from a BoostLine object and a DataFrame.
Line attribute data is transformed into edge attribute data, and an extra 'length' edge attribute with
Euclidean or geodesic (see attribute \code{lonlat}) line lengths is added.
}

