% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boost_lines.R
\name{bSplit}
\alias{bSplit}
\title{Split Lines}
\usage{
bSplit(bl, maxlen)
}
\arguments{
\item{bl}{A BoostLines object.}

\item{maxlen}{The maximum length of the split lines.}
}
\value{
A list of (1) a new, split BoostLines object and (2) an integer vector of line IDs.
}
\description{
Splits lines into shorter lines of length maxlen.
}
\details{
Input lines are split into shorter lines of length maxlen, plus a remainder of length smaller than maxlength.
Input lines shorter than maxlen are left untouched.
}

