% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boost_lines.R
\name{remove_multilines}
\alias{remove_multilines}
\title{Eliminate Multilines}
\usage{
remove_multilines(s)
}
\arguments{
\item{s}{A SpatialLinesDataFrame or SpatialLines object.}
}
\value{
A SpatialLinesDataFrame or SpatialLines object.
}
\description{
Removes all multilines from a SpatialLinesDataFrame by replacing them with simple lines.
Attribute data for multilines is duplicated accordingly.
}

