% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boost_lines.R
\name{unboost}
\alias{unboost}
\title{Cast BoostLines as SpatialLines}
\usage{
unboost(bl)
}
\arguments{
\item{bl}{A BoostLines object.}
}
\value{
A SpatialLines object.
}
\description{
Creates a SpatialLines object.
}

